Crate tokio_ecosystem_tests

tokio-ecosystem-tests

set of tests and notes about tokio.

copy or loop

definitions

AsyncRead represents any source that contains bytes (e.g. &[u8])

AsyncWrite represents any destination that can hold bytes (e.g. Vec<u8>)

StreamExt provides stream.next() method to allow for iteration in a while loop

(complementary yet different from futures_util::stream::StreamExt)

utility methods & structs

copy comes in handy to copy bytes from a AsyncRead to a buffer AsyncWrite

StreamReader comes in handy to turn a Stream::<Bytes> into a AsyncRead

iter provides a convenient way to turn an Iterator into a Stream::<Bytes>

manual iteration vs utility method

directly iterating over chunks of bytes is preferable to using copy which uses an intermediate buffer, but one can look at copy_buf to achieve the same result

see benchmark

examples

//! illustrate correlated usages of [`copy`] and [`StreamReader`]
//!
//! [`copy`]: https://docs.rs/tokio/latest/tokio/io/fn.copy.html
//! [`StreamReader`]: https://docs.rs/tokio-util/latest/tokio_util/io/struct.StreamReader.html

#[cfg(test)]
mod tests {
    use bytes::{BufMut, Bytes};
    use tokio;
    use tokio::io;
    use tokio_stream::StreamExt;
    use tokio_util::io::StreamReader;

    #[tokio::test]
    async fn tokio_doc_fn_copy() -> io::Result<()> {
        let mut reader: &[u8] = b"hello";
        let mut writer: Vec<u8> = vec![];

        io::copy(&mut reader, &mut writer).await?;

        assert_eq!(&b"hello"[..], &writer[..]);
        Ok(())
    }

    #[allow(non_snake_case)]
    #[tokio::test]
    async fn struct_StreamReader() -> io::Result<()> {
        let stream = tokio_stream::iter(vec![
            io::Result::Ok(Bytes::from_static(&[0, 1, 2, 3])),
            io::Result::Ok(Bytes::from_static(&[4, 5, 6, 7])),
            io::Result::Ok(Bytes::from_static(&[8, 9, 10, 11])),
        ]);

        let mut reader = StreamReader::new(stream);
        let mut writer: Vec<u8> = vec![];

        // note that the reader is consumed after copy (and left empty)
        io::copy(&mut reader, &mut writer).await?;

        assert_eq!(&[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11], &writer[..]);
        Ok(())
    }

    #[tokio::test]
    async fn loop_over_chunks() -> io::Result<()> {
        let mut stream = tokio_stream::iter(vec![
            io::Result::Ok(Bytes::from_static(&[0, 1, 2, 3])),
            io::Result::Ok(Bytes::from_static(&[4, 5, 6, 7])),
            io::Result::Ok(Bytes::from_static(&[8, 9, 10, 11])),
        ]);

        let mut buf = Vec::<u8>::new();
        while let Some(io::Result::Ok(bytes)) = stream.next().await {
            buf.put(bytes);
        }

        assert_eq!(&[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11], &buf[..]);
        Ok(())
    }
}

range

definitions

stream! returns an anonymous type implementing the Stream trait

stream transformations

create dynamic streams from input, map stream’s output, chain different streams

examples

//! illustrate usage of [`stream!`]
//!
//! [`stream!`]: https://docs.rs/async-stream/latest/async_stream/macro.stream.html

use std::ops::RangeInclusive;

use async_stream::stream;
use tokio::io;
use tokio_stream::Stream;

pub fn from_range(range: RangeInclusive<u32>) -> impl Stream<Item = io::Result<u32>> {
    stream! {
        for i in range {
            yield io::Result::Ok(i); // notice here that it yields a tokio::io::Result
        }
    }
}

#[cfg(test)]
mod tests {
    use super::from_range;
    use async_stream::stream;
    use futures::pin_mut;
    use tokio::io;
    use tokio_stream::StreamExt;

    #[tokio::test]
    async fn tokio_doc_macro_stream() -> io::Result<()> {
        let first = stream! {
            for i in 0..=3 {
                yield i; // notice here that it yields the number directly
            }
        };
        let second = from_range(4..=7);
        let third = from_range(8..=11);

        pin_mut!(first);
        pin_mut!(second);
        pin_mut!(third);

        let mut stream = first
            // map first digits into tokio::io::Result
            .map(io::Result::Ok)
            // subsequent streams already yield tokio::io::Result
            .chain(second)
            .chain(third);

        let mut buf = Vec::<u32>::new();
        while let Some(io::Result::Ok(item)) = stream.next().await {
            buf.push(item);
        }

        assert_eq!(&[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11], &buf[..]);
        Ok(())
    }
}

StreamExt(s)

definitions

stream transformations

although differents, both StreamExt from tokio and futures crates can be mixed together

examples

example below slightly elaborate over the StreamExt trait example

//! illustrate over documentation usage for [`tokio_stream::StreamExt`] and [`futures::stream::StreamExt`]
//!
//! [`tokio_stream::StreamExt`]: https://docs.rs/tokio-stream/latest/tokio_stream/trait.StreamExt.html
//! [`futures::stream::StreamExt`]: https://docs.rs/futures/latest/futures/stream/trait.StreamExt.html

#[cfg(test)]
mod tests {
    use futures::stream::StreamExt;
    use tokio;
    use tokio::io;

    #[allow(non_snake_case)]
    #[tokio::test]
    async fn tokio_doc_trait_StreamExt() -> io::Result<()> {
        let a = tokio_stream::iter(vec![1, 3, 5]);
        let b = tokio_stream::iter(vec![2, 4, 6]);

        let merged = tokio_stream::StreamExt::merge(a, b);
        let output: Vec<_> = merged.collect().await;

        assert_eq!(output, vec![1, 2, 3, 4, 5, 6]);
        Ok(())
    }

    #[tokio::test]
    async fn mixing_both_methods() -> io::Result<()> {
        let a = tokio_stream::iter(vec![1, 3, 5, 7, 9, 11]);
        let b = tokio_stream::iter(vec![2, 4, 6, 8, 10, 12]);

        let output = tokio_stream::StreamExt::merge(a, b);
        let output = output.filter(|x| futures::future::ready(x % 3 == 0));
        let output = tokio_stream::StreamExt::filter(output, |x| x % 2 == 0);
        let output = output.filter(|x| futures::future::ready(x % 4 == 0));

        let output: Vec<_> = output.collect().await;
        assert_eq!(output, vec![12]);
        Ok(())
    }
}

Modules

copy_or_loop

illustrate correlated usages of copy and StreamReader

range

illustrate usage of stream!

streamext

illustrate over documentation usage for tokio_stream::StreamExt and futures::stream::StreamExt